pyproject.toml
[build-system]
requires = ["setuptools >= 68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "luma"
version = "0.1.0"
requires-python = ">= 3.12"
dependencies = ["anthropic", "pydantic", "python-dotenv", "logfire"]

[project.optional-dependencies]
dev = ["pytest", "pexpect", "pydantic-evals[logfire]"]

[project.scripts]
luma = "luma.cli:main"

[tool.setuptools.package-data]
"luma.agent" = ["prompts/*.md"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
